requirements.txt
flask
tensorflow
numpy
pillow
split-folders
